Parallelize load_profiles

+++ backend/app/load_data_helpers/get_games_files.py
@@ -18,18 +18,18 @@ extract_date_re = re.compile(
 def download_game(game_url):
     """
     Downloads the file at the given URL to /tmp/filename
-    and returns the local filepath.
+    and returns local_filepath, already_downloaded
     """
     os.makedirs("/tmp", exist_ok=True)
     filepath = "/tmp/" + game_url.split("/")[-1]
     if os.path.exists(filepath):
-        return filepath
+        return filepath, True
     with requests.get(game_url, stream=True) as r:
         file_size = int(r.headers.get("Content-Length", 0)) or None
         with open(filepath, "wb") as f:
             with tqdm.wrapattr(r.raw, "read", total=file_size) as r_raw:
                 shutil.copyfileobj(r_raw, f)
-    return filepath
+    return filepath, False
 
 
 def get_games_files(
@@ -67,13 +67,13 @@ def get_games_files(
             break
 
         print(f"Downloading game file for {date}...")
-        filepath = download_game(game_url)
+        filepath, already_downloaded = download_game(game_url)
 
         @contextlib.contextmanager
         def get_context():
             with bz2.open(filepath, "rt") as f:
                 yield f
-            if not save_files:
+            if not save_files and not already_downloaded:
                 os.remove(filepath)
 
         yield get_context
